# lib/api-zod/src/generated/types/newsItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NewsItemChangeStatus } from './newsItemChangeStatus';
import type { NewsItemClassification } from './newsItemClassification';
import type { NewsItemExtractionConfidence } from './newsItemExtractionConfidence';
import type { NewsItemSource } from './newsItemSource';

export interface NewsItem {
  /** Stable URL-safe identifier for the item-detail endpoint */
  id: string;
  title: string;
  link: string;
  summary: string;
  publishedDate: Date;
  /** @nullable */
  revisedDate?: Date | null;
  /** Best-available literal timestamp (date + time) of when the item was posted at the source */
  postedAt: Date;
  publishedLabel: string;
  categories: string[];
  monthYear: string;
  /**
     * Full HTML body of the article; null for bulletins
     * @nullable
     */
  body?: string | null;
  source: NewsItemSource;
  classification: NewsItemClassification;
  /** HP internal category tags derived from DHCS baseline mapping plus content analysis */
  hpMappings: string[];
  /** Inferred from explicit source wording only; empty when no bounded evidence is found */
  affectedEntities: string[];
  /** Inferred from explicit action language in source text; not a DHCS-authored determination */
  actionRequired: boolean;
  /**
     * Literal source sentence supporting actionRequired; null when uncertain
     * @nullable
     */
  actionSummary: string | null;
  /**
     * Literal source date following an explicit effective-date cue; null when uncertain
     * @nullable
     */
  effectiveDate: string | null;
  /**
     * Literal source date following an explicit deadline cue; null when uncertain
     * @nullable
     */
  deadlineDate: string | null;
  /** Explicit program or coding terms from source text only */
  keyPrograms: string[];
  /** revised only when revisedDate is later than publishedDate; safe for future recently-revised filtering */
  changeStatus: NewsItemChangeStatus;
  /** Confidence in deterministic, source-bounded inferred metadata; not model confidence */
  extractionConfidence: NewsItemExtractionConfidence;
}